Interview\AbstractAnswer: introduced to some commands KP-8034

+++ src/Core/SharedKernels/DataCollection/DataCollection/Implementation/Aggregates/Interview.MultipleOptionsQuestion.cs
@@ -7,7 +7,7 @@ namespace WB.Core.SharedKernels.DataCollection.Implementation.Aggregates
 {
     public partial class Interview
     {
-        public void AnswerMultipleOptionsQuestion(Guid userId, Guid questionId, RosterVector rosterVector, DateTime answerTime, decimal[] selectedValues)
+        public void AnswerMultipleOptionsQuestion(Guid userId, Guid questionId, RosterVector rosterVector, DateTime answerTime, int[] selectedValues)
         {
             new InterviewPropertiesInvariants(this.properties).RequireAnswerCanBeChanged();
